define the one and only internal version of exists, use for field, move fieldvar
