+++ src/lib/bool.ts
@@ -11,8 +11,9 @@ import { Bool as B } from '../provable/field-bigint.js';
 import { defineBinable } from '../bindings/lib/binable.js';
 import { NonNegativeInteger } from '../bindings/crypto/non-negative.js';
 import { asProver } from './provable-context.js';
-import { existsOne } from './gadgets/common.js';
+import { existsOne } from './provable-core/exists.js';
 import { assertMul } from './gadgets/compatible.js';
+import { setBoolConstructor } from './provable-core/field-constructor.js';
 
 export { BoolVar, Bool };
 
@@ -365,6 +366,7 @@ class Bool {
     },
   };
 }
+setBoolConstructor(Bool);
 
 const BoolBinable = defineBinable({
   toBytes(b: Bool) {
